Switch to custom Instructions wrapper type
This switches to a instructions wrapper type which advances a head
pointer, rather than consuming the vector. this is both more efficient,
and gets us ready for a refactor. The way an 8086 works, is that it will
essentially decode and execute as it goes along, rather than decoding
all at once and then executing all at once. This means that if it gets a
jump instruction, it needs to use the instruction pointer (ip register)
to then jump forward or back a specific number of bytes. When it does
that it then starts decoding again from that position (rather than
re-using previously decoded data)

The point of this repo is to learn how a CPU actually works - so it
should model one as closely as possible. The next step will be to model
a streaming approach, rather than decode and then execute approach.