Encode loop and add func to spit out mem as file
This adds handling of the loop operation, along with a couple new fun
listings that draw gradient squares. I added a binary file so we can
take a binary file in, and accept an output file to dump the "memory" we
are simulating. This can then be fed into gimp to interpet the raw bytes
as image data, and it worked. Fun!